Breaking StorageAdapter implementations into their own class files
