+++ valkyrie/lib/valkyrie.rb
@@ -34,6 +34,7 @@ module Valkyrie
   require 'valkyrie/decorators/decorator_with_arguments'
   require 'valkyrie/model/access_controls'
   require 'valkyrie/indexers/access_controls_indexer'
+  require 'valkyrie/storage'
   require 'valkyrie/vocab/pcdm_use'
   require 'generators/valkyrie/model_generator'
   require 'valkyrie/engine' if defined?(Rails)
