+++ valkyrie/lib/valkyrie/storage.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+module Valkyrie
+  module Storage
+    require 'valkyrie/storage/disk'
+    require 'valkyrie/storage/fedora'
+    require 'valkyrie/storage/memory'
+  end
+end
