+++ valkyrie/spec/valkyrie/storage/disk_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 require 'valkyrie/specs/shared_specs'
 include ActionDispatch::TestProcess
 
-RSpec.describe Valkyrie::StorageAdapter::Disk do
+RSpec.describe Valkyrie::Storage::Disk do
   it_behaves_like "a Valkyrie::StorageAdapter"
   let(:storage_adapter) { described_class.new(base_path: ROOT_PATH.join("tmp", "files_test")) }
   let(:file) { fixture_file_upload('files/example.tif', 'image/tiff') }
