+++ valkyrie/spec/valkyrie/storage/memory_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 require 'valkyrie/specs/shared_specs'
 include ActionDispatch::TestProcess
 
-RSpec.describe Valkyrie::StorageAdapter::Memory do
+RSpec.describe Valkyrie::Storage::Memory do
   it_behaves_like "a Valkyrie::StorageAdapter"
   let(:storage_adapter) { described_class.new }
   let(:file) { fixture_file_upload('files/example.tif', 'image/tiff') }
